moved definition of metainfo interface
