+++ src/index.tsx
@@ -8,6 +8,7 @@ import { rootReducer } from './store/store';
 import { createStore } from 'redux';
 import { composeWithDevTools } from 'redux-devtools-extension';
 
+// https://medium.com/@zalmoxis/using-redux-devtools-in-production-4c5b56c5600f
 const composeEnhancers = composeWithDevTools({
   // options like actionSanitizer, stateSanitizer
 });
